+++ examples/examples_go_test.go
@@ -14,14 +14,26 @@
 package examples
 
 import (
+	"path/filepath"
+	"testing"
+
 	"github.com/pulumi/pulumi/pkg/v3/testing/integration"
 )
+
+func TestAccRealmGo(t *testing.T) {
+	test := getGoBaseOptions().
+		With(integration.ProgramTestOptions{
+			Dir: filepath.Join(getCwd(t), "realm", "go"),
+		})
+
+	integration.ProgramTest(t, &test)
+}
 
 func getGoBaseOptions() integration.ProgramTestOptions {
 	base := getBaseOptions()
 	goBase := base.With(integration.ProgramTestOptions{
 		Dependencies: []string{
-			"github.com/pulumi/pulumi-keycloak/sdk/v2",
+			"github.com/pulumi/pulumi-keycloak/sdk/v4",
 		},
 	})